+++ bin/describe.ps1
@@ -24,7 +24,7 @@ param(
 . "$PSScriptRoot\..\lib\manifest.ps1"
 . "$PSScriptRoot\..\lib\description.ps1"
 
-$Dir = Resolve-Path $Dir
+$Dir = Convert-Path $Dir
 $Queue = @()
 
 Get-ChildItem $Dir "$App.json" | ForEach-Object {